feat: 新增订单状态别名方法及快递订单服务接口
refactor(OrderMain): 添加getOrderStatus方法作为status的别名
feat(ExpressOrderService): 新增快递订单相关服务接口
fix: 修复OrderMain中status字段的更新逻辑
style: 优化代码格式和变量命名
docs: 更新相关SQL字段注释和长度限制

+++ youyinda-backend/src/main/java/com/youyinda/task/ThirdApiSyncTask.java
@@ -71,8 +71,8 @@ public class ThirdApiSyncTask {
         );
 
         if (existPrice != null) {
-            existPrice.setBasePrice(price.getBasePrice());
-            existPrice.setMinOrderPrice(price.getMinOrderPrice());
+            existPrice.setBasePrice(price.getBasePrice().doubleValue());
+            existPrice.setMinOrderPrice(price.getMinOrderPrice().doubleValue());
             printBasePriceService.updateById(existPrice);
         } else {
             PrintBasePrice newPrice = new PrintBasePrice();
@@ -80,8 +80,8 @@ public class ThirdApiSyncTask {
             newPrice.setPaperType(price.getPaperType());
             newPrice.setColorType(price.getColorType());
             newPrice.setSingleDouble(price.getSingleDouble());
-            newPrice.setBasePrice(price.getBasePrice());
-            newPrice.setMinOrderPrice(price.getMinOrderPrice());
+            newPrice.setBasePrice(price.getBasePrice().doubleValue());
+            newPrice.setMinOrderPrice(price.getMinOrderPrice().doubleValue());
             newPrice.setStatus(1);
             printBasePriceService.save(newPrice);
         }
